+++ src/studio_sdk/types/account.py
@@ -10,6 +10,9 @@ class Account(BaseModel):
     account_id: str
     """Account ID for the account."""
 
+    account_number: str
+    """Account number for the account."""
+
     entity_id: str
     """Entity ID for the legal entity."""