+++ src/studio_sdk/types/shared/trade.py
@@ -33,5 +33,8 @@ class Trade(BaseModel):
     account_id: Optional[str] = None
     """Account ID for the account."""
 
+    account_number: Optional[str] = None
+    """Account number for the account."""
+
     symbol: Optional[str] = None
     """The symbol this trade was for."""