+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.1.0-alpha.15 (2024-12-09)
+
+Full Changelog: [v0.1.0-alpha.14...v0.1.0-alpha.15](https://github.com/clear-street/studio-sdk-python/compare/v0.1.0-alpha.14...v0.1.0-alpha.15)
+
+### Features
+
+* **api:** account number support ([#74](https://github.com/clear-street/studio-sdk-python/issues/74)) ([c91236d](https://github.com/clear-street/studio-sdk-python/commit/c91236de909214b523701b83e10eb3366a7ec426))
+
+
+### Chores
+
+* **internal:** bump pyright ([#71](https://github.com/clear-street/studio-sdk-python/issues/71)) ([75e4f25](https://github.com/clear-street/studio-sdk-python/commit/75e4f25fcca33bae792841c978ea66c29b8db61a))
+
 ## 0.1.0-alpha.14 (2024-11-29)
 
 Full Changelog: [v0.1.0-alpha.13...v0.1.0-alpha.14](https://github.com/clear-street/studio-sdk-python/compare/v0.1.0-alpha.13...v0.1.0-alpha.14)